requirements.txt
pywebview
PyInstaller
pillow
